fix size_t and int warnings

+++ src/IO/H5/VolumeData.hpp
@@ -259,15 +259,15 @@ Mesh<Dim> mesh_for_grid(
     const std::vector<std::vector<Spectral::Quadrature>>& all_quadratures);
 
 template <size_t SpatialDim>
-std::pair<std::vector<std::array<int, SpatialDim>>,
-          std::vector<std::array<int, SpatialDim>>>
+std::pair<std::vector<std::array<size_t, SpatialDim>>,
+          std::vector<std::array<size_t, SpatialDim>>>
 compute_element_refinements_and_indices(
     const std::vector<std::string>& block_grid_names);
 
 template <size_t SpatialDim>
 std::vector<std::array<SegmentId, SpatialDim>> create_SegmentIds(
-    const std::pair<std::vector<std::array<int, SpatialDim>>,
-                    std::vector<std::array<int, SpatialDim>>>&
+    const std::pair<std::vector<std::array<size_t, SpatialDim>>,
+                    std::vector<std::array<size_t, SpatialDim>>>&
         refinements_and_indices);
 
 }  // namespace h5
